+++ src/Tests/Integration/SupportedModels-Checks.Tests.ps1
@@ -27,6 +27,7 @@ Describe 'Supported Models Checks' -Tag Integration {
             'anthropic.claude-3-sonnet-20240229-v1:0'
             'anthropic.claude-3-5-sonnet-20241022-v2:0',
             'anthropic.claude-3-5-sonnet-20240620-v1:0'
+            'anthropic.claude-3-7-sonnet-20250219-v1:0'
             # 'anthropic.claude-instant-v1'
             'cohere.command-text-v14'
             'cohere.command-light-text-v14'