+++ src/pwshBedrock/Public/Invoke-ConverseAPI.ps1
@@ -276,6 +276,7 @@ function Invoke-ConverseAPI {
             'anthropic.claude-3-sonnet-20240229-v1:0',
             'anthropic.claude-3-5-sonnet-20241022-v2:0',
             'anthropic.claude-3-5-sonnet-20240620-v1:0',
+            'anthropic.claude-3-7-sonnet-20250219-v1:0',
             # 'cohere.command-text-v14', # *note: not supported by Converse API
             # 'cohere.command-light-text-v14', # *note: not supported by Converse API
             'cohere.command-r-v1:0',