+++ src/pwshBedrock/Public/Save-ModelContext.ps1
@@ -54,6 +54,7 @@ function Save-ModelContext {
             'anthropic.claude-3-opus-20240229-v1:0',
             'anthropic.claude-3-5-sonnet-20241022-v2:0',
             'anthropic.claude-3-5-sonnet-20240620-v1:0',
+            'anthropic.claude-3-7-sonnet-20250219-v1:0',
             # 'cohere.command-text-v14',
             # 'cohere.command-light-text-v14',
             'cohere.command-r-v1:0',